Add golangci-lint and fix code quality issues
Add comprehensive golangci-lint configuration with linters for
detecting IDE-level issues: unparam, unconvert, gocritic, dupl,
gocyclo, and misspell.

- Add .golangci.yml with optimized thresholds
- Fix unparam: remove always-constant function parameters
- Fix ifElseChain: convert to switch statements across handlers
- Fix unlambda: use direct function references in render.go
- Fix emptyStringTest: use s == "" instead of len(s) == 0
- Fix assignOp: use += operator where applicable
- Fix appendAssign: proper slice append patterns
- Remove redundant type conversions and unused test code

+++ modules/migrator/sources/elefant/importer_test.go
@@ -146,7 +146,7 @@ func TestImportUsers_Success(t *testing.T) {
 			t.Fatalf("failed to create user: %v", err)
 		}
 		result.UsersImported++
-		_ = tracker.TrackImportedItem(ctx, s.Name(), "user", int64(u.ID))
+		_ = tracker.TrackImportedItem(ctx, s.Name(), "user", u.ID)
 	}
 
 	// Verify results
